// src/app/models/renpol.ts
export interface Renpol {
    fecha: string;
    tda: string;
    idcli: number;
    numcli: string;
    nombre: string;
    sino: string;
    concep: string;
    tipo: string;
    bonificacion: number;
    recargo: number;
    importe: number;
    salcli: number;
    moneda: string;
    fecven:string;
    comis: number;
    dias: number;
    cobr2: string;
    usuario:string;
    ace:string;
    conse:number;
    idrenpol:  number;
    idpoliza: number;
    idcobra:  number;
    idusuario: number;
    foliobon:  number;
    iduuid: number;
    idfacfon: number;
    feccap:string;
    uuid: string;

}